Put tests in a single file
