--- .eslintrc.json
@@ -1,9 +1,0 @@
-{
-  "extends": "standard",
-  "parser": "babel-eslint",
-  "rules": {
-    "yoda": 0,
-    "semi": [2, "never"],
-    "no-extra-semi": 2
-  }
-}
